backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-agent-evaluation-backend"
version = "0.1.0"
description = "AI Agent Evaluation Pipeline - Backend API"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = []
keywords = ["fastapi", "evaluation", "ai", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy>=2.0.0",
    "redis>=5.0.0",
    "celery[redis]>=5.3.0",
    "openai>=1.12.0",
    "litellm>=1.30.0",
    "httpx>=0.26.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[project.scripts]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
